Fixed the polling bug where only 1 node's frame is recorded.
The bug was found in the developed function get_one_frame ( similar to get_all_frames ).
An internal counter resets on each call of the function causing the function to write only on the first element recursively. All subsequent calls to the function writes on the first element, thus causing the bug.